libpng 1.6 workaround

+++ inc/image.php
@@ -361,8 +361,10 @@ class ImageConvert extends ImageBase {
 					$this->width,
 					$this->height,
 					escapeshellarg($this->temp)))) || !file_exists($this->temp)) {
-				$this->destroy();
-				error('Failed to resize image!', null, $error);
+					if (strpos($error, "known incorrect sRGB profile") === "false") {
+						$this->destroy();
+						error('Failed to resize image!', null, $error);
+					}
 			}
 			if ($size = $this->get_size($this->temp)) {
 				$this->width = $size[0];
